# GYMaster-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GYMaster",
    "version": "v1"
  },
  "paths": {
...
    "/api/Gimnasios/Validar": {
      "post": {
        "tags": [
          "Gimnasios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CredencialesDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Gimnasios/Buscar": {
      "get": {
        "tags": [
          "Gimnasios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GimnasiosDTO"
            }
          }
        }
      }
    },
    "/api/Gimnasios/Registrar": {
      "post": {
        "tags": [
          "Gimnasios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GimnasiosDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Gimnasios/Modificar": {
      "put": {
        "tags": [
          "Gimnasios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GimnasiosDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Gimnasios/Actualizar": {
      "patch": {
        "tags": [
          "Gimnasios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CredencialesDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Membresias/Mostrar": {
      "get": {
        "tags": [
          "Membresias"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MembresiasDTO"
              }
            }
          }
        }
      }
    },
    "/api/Membresias/Catalogo": {
      "get": {
        "tags": [
          "Membresias"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MembresiasDTO"
              }
            }
          }
        }
      }
    },
    "/api/Membresias/Registrar": {
      "post": {
        "tags": [
          "Membresias"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MembresiasDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Membresias/Modificar": {
      "put": {
        "tags": [
          "Membresias"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MembresiasDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Paypal/PaymentWithPaypal": {
      "get": {
        "tags": [
          "Paypal"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Cancel",
            "type": "string"
          },
          {
            "in": "query",
            "name": "guid",
            "type": "string",
            "default": ""
          },
          {
            "in": "query",
            "name": "paymentId",
            "type": "string",
            "default": ""
          },
          {
            "in": "query",
            "name": "token",
            "type": "string",
            "default": ""
          },
          {
            "in": "query",
            "name": "PayerID",
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CredencialesDTO": {
      "type": "object",
      "properties": {
        "usuario": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        }
      }
    },
...
    "GimnasiosDTO": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        }
      }
    },
...
    "MembresiasDTO": {
      "type": "object",
      "properties": {
        "usuarioGimnasios": {
          "type": "string"
        },
        "tipo": {
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "periodo": {
          "format": "int32",
          "type": "integer"
        },
        "idEstatus": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}